requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pytest>=8.0.0
urllib3<2
